v0.6
- New levels
- New game features
- Better code arrange

+++ LevelObjs.h
@@ -5,7 +5,7 @@
 #include <vector>
 #include "Enums.h"
 
-extern const int TILE_SIZE;
+extern const float TILE_SIZE;
 
 class Block;
 class PushableBlock;
